DeleteFiles plugin
