+++ website/src/examples/dashboard/app.es6
@@ -5,6 +5,7 @@ const Dropbox = require('uppy/lib/plugins/Dropbox')
 const Instagram = require('uppy/lib/plugins/Instagram')
 const Webcam = require('uppy/lib/plugins/Webcam')
 const Tus = require('uppy/lib/plugins/Tus')
+const DeleteFiles = require('uppy/lib/plugins/DeleteFiles')
 
 const UPPY_SERVER = require('../env')
 
@@ -32,7 +33,6 @@ function uppyInit () {
 
   const uppy = Uppy({
     debug: true,
-    removeAfterUpload: opts.removeAfterUpload,
     autoProceed: opts.autoProceed,
     restrictions: opts.restrictions ? restrictions : ''
   })
@@ -67,6 +67,11 @@ function uppyInit () {
   }
 
   uppy.use(Tus, { endpoint: TUS_ENDPOINT, resume: true })
+
+  if (opts.removeAfterUpload) {
+    uppy.use(DeleteFiles, { endpoint: '//api2.transloadit.com' })
+  }
+
   uppy.run()
 
   uppy.on('complete', result => {
